+++ src/Exceptions/Solutions/Providers/MissingACLWildcardsTraitSolutionProvider.php
@@ -4,8 +4,8 @@ namespace Junges\ACL\Exceptions\Solutions\Providers;
 
 use Facade\IgnitionContracts\BaseSolution;
 use Facade\IgnitionContracts\HasSolutionsForThrowable;
+use Junges\ACL\Concerns\ACLWildcardsTrait;
 use Junges\ACL\Exceptions\Solutions\AddMissingACLWildcardsTraitSolution;
-use Junges\ACL\Traits\ACLWildcardsTrait;
 use ReflectionClass;
 use Throwable;