+++ src/Exceptions/Solutions/Providers/MissingGroupsTraitSolutionProvider.php
@@ -4,8 +4,8 @@ namespace Junges\ACL\Exceptions\Solutions\Providers;
 
 use Facade\IgnitionContracts\BaseSolution;
 use Facade\IgnitionContracts\HasSolutionsForThrowable;
+use Junges\ACL\Concerns\GroupsTrait;
 use Junges\ACL\Exceptions\Solutions\AddMissingGroupsTraitSolution;
-use Junges\ACL\Traits\GroupsTrait;
 use ReflectionClass;
 use Throwable;