+++ tests/Permission.php
@@ -4,12 +4,13 @@ namespace Junges\ACL\Tests;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Junges\ACL\Concerns\PermissionsTrait;
 use Junges\ACL\Events\PermissionSaving;
-use Junges\ACL\Traits\PermissionsTrait;
 
 class Permission extends Model
 {
-    use SoftDeletes, PermissionsTrait;
+    use SoftDeletes;
+    use PermissionsTrait;
 
     protected $table = 'test_permissions';